+++ src/fhapi/system.h
@@ -1,0 +1,13 @@
+/*
+#   clove
+#
+#   Copyright (C) 2016-2025 Muresan Vlad
+#
+#   This project is free software; you can redistribute it and/or modify it
+#   under the terms of the MIT license. See LICENSE.md for details.
+*/
+#pragma once
+
+#include "../3rdparty/FH/src/fh.h"
+
+void fh_system_register(struct fh_program *prog);